refactor(api): split friend-connection.service.ts into services/friend-connection/ (api-drgh)
