+++ apps/api/src/__tests__/routes/account/friends/accept-reject.test.ts
@@ -12,16 +12,22 @@ import type { ApiErrorResponse } from "@pluralscape/types";
 
 // ── Mocks ────────────────────────────────────────────────────────
 
-vi.mock("../../../../services/friend-connection.service.js", () => ({
-  listFriendConnections: vi.fn(),
-  getFriendConnection: vi.fn(),
-  blockFriendConnection: vi.fn(),
-  removeFriendConnection: vi.fn(),
-  acceptFriendConnection: vi.fn(),
-  rejectFriendConnection: vi.fn(),
-  updateFriendVisibility: vi.fn(),
+vi.mock("../../../../services/friend-connection/lifecycle.js", () => ({
   archiveFriendConnection: vi.fn(),
   restoreFriendConnection: vi.fn(),
+}));
+vi.mock("../../../../services/friend-connection/queries.js", () => ({
+  getFriendConnection: vi.fn(),
+  listFriendConnections: vi.fn(),
+}));
+vi.mock("../../../../services/friend-connection/transitions.js", () => ({
+  acceptFriendConnection: vi.fn(),
+  blockFriendConnection: vi.fn(),
+  rejectFriendConnection: vi.fn(),
+  removeFriendConnection: vi.fn(),
+}));
+vi.mock("../../../../services/friend-connection/update.js", () => ({
+  updateFriendVisibility: vi.fn(),
 }));
 
 vi.mock("../../../../lib/audit-writer.js", () => mockAuditWriterFactory());
@@ -34,7 +40,7 @@ vi.mock("../../../../middleware/auth.js", () => mockAccountOnlyAuthFactory());
 // ── Imports after mocks ──────────────────────────────────────────
 
 const { acceptFriendConnection, rejectFriendConnection } =
-  await import("../../../../services/friend-connection.service.js");
+  await import("../../../../services/friend-connection/transitions.js");
 const { createAuditWriter } = await import("../../../../lib/audit-writer.js");
 const { accountRoutes } = await import("../../../../routes/account/index.js");
