+++ apps/api/src/__tests__/routes/account/friends/key-grants.test.ts
@@ -16,15 +16,21 @@ vi.mock("../../../../services/key-grant.service.js", () => ({
   listReceivedKeyGrants: vi.fn(),
 }));
 
-vi.mock("../../../../services/friend-connection.service.js", () => ({
-  acceptFriendConnection: vi.fn(),
+vi.mock("../../../../services/friend-connection/lifecycle.js", () => ({
   archiveFriendConnection: vi.fn(),
-  blockFriendConnection: vi.fn(),
+  restoreFriendConnection: vi.fn(),
+}));
+vi.mock("../../../../services/friend-connection/queries.js", () => ({
   getFriendConnection: vi.fn(),
   listFriendConnections: vi.fn(),
+}));
+vi.mock("../../../../services/friend-connection/transitions.js", () => ({
+  acceptFriendConnection: vi.fn(),
+  blockFriendConnection: vi.fn(),
   rejectFriendConnection: vi.fn(),
   removeFriendConnection: vi.fn(),
-  restoreFriendConnection: vi.fn(),
+}));
+vi.mock("../../../../services/friend-connection/update.js", () => ({
   updateFriendVisibility: vi.fn(),
 }));
