+++ apps/api/src/__tests__/routes/account/pin.test.ts
@@ -47,15 +47,21 @@ vi.mock("../../../services/friend-code.service.js", () => ({
   revokeFriendCode: vi.fn(),
 }));
 
-vi.mock("../../../services/friend-connection.service.js", () => ({
-  acceptFriendConnection: vi.fn(),
+vi.mock("../../../services/friend-connection/lifecycle.js", () => ({
   archiveFriendConnection: vi.fn(),
-  blockFriendConnection: vi.fn(),
+  restoreFriendConnection: vi.fn(),
+}));
+vi.mock("../../../services/friend-connection/queries.js", () => ({
   getFriendConnection: vi.fn(),
   listFriendConnections: vi.fn(),
+}));
+vi.mock("../../../services/friend-connection/transitions.js", () => ({
+  acceptFriendConnection: vi.fn(),
+  blockFriendConnection: vi.fn(),
   rejectFriendConnection: vi.fn(),
   removeFriendConnection: vi.fn(),
-  restoreFriendConnection: vi.fn(),
+}));
+vi.mock("../../../services/friend-connection/update.js", () => ({
   updateFriendVisibility: vi.fn(),
 }));
