+++ apps/api/src/routes/account/friends/accept.ts
@@ -1,4 +1,4 @@
-import { acceptFriendConnection } from "../../../services/friend-connection.service.js";
+import { acceptFriendConnection } from "../../../services/friend-connection/transitions.js";
 
 import { createFriendActionRoute } from "./create-friend-action-route.js";
