+++ apps/api/src/routes/account/friends/get.ts
@@ -5,7 +5,7 @@ import { getDb } from "../../../lib/db.js";
 import { requireIdParam } from "../../../lib/id-param.js";
 import { envelope } from "../../../lib/response.js";
 import { createCategoryRateLimiter } from "../../../middleware/rate-limit.js";
-import { getFriendConnection } from "../../../services/friend-connection.service.js";
+import { getFriendConnection } from "../../../services/friend-connection/queries.js";
 
 import type { AuthEnv } from "../../../lib/auth-context.js";
