+++ apps/api/src/routes/account/friends/remove.ts
@@ -6,7 +6,7 @@ import { getDb } from "../../../lib/db.js";
 import { requireIdParam } from "../../../lib/id-param.js";
 import { envelope } from "../../../lib/response.js";
 import { createCategoryRateLimiter } from "../../../middleware/rate-limit.js";
-import { removeFriendConnection } from "../../../services/friend-connection.service.js";
+import { removeFriendConnection } from "../../../services/friend-connection/transitions.js";
 
 import type { AuthEnv } from "../../../lib/auth-context.js";
