+++ apps/api/src/routes/account/friends/restore.ts
@@ -6,7 +6,7 @@ import { getDb } from "../../../lib/db.js";
 import { requireIdParam } from "../../../lib/id-param.js";
 import { envelope } from "../../../lib/response.js";
 import { createCategoryRateLimiter } from "../../../middleware/rate-limit.js";
-import { restoreFriendConnection } from "../../../services/friend-connection.service.js";
+import { restoreFriendConnection } from "../../../services/friend-connection/lifecycle.js";
 
 import type { AuthEnv } from "../../../lib/auth-context.js";
