+++ apps/api/src/routes/account/friends/visibility.ts
@@ -10,7 +10,7 @@ import { requireIdParam } from "../../../lib/id-param.js";
 import { parseJsonBody } from "../../../lib/parse-json-body.js";
 import { envelope } from "../../../lib/response.js";
 import { createCategoryRateLimiter } from "../../../middleware/rate-limit.js";
-import { updateFriendVisibility } from "../../../services/friend-connection.service.js";
+import { updateFriendVisibility } from "../../../services/friend-connection/update.js";
 
 import type { AuthEnv } from "../../../lib/auth-context.js";
