+++ apps/api/src/services/friend-connection/transitions.ts
@@ -1,73 +1,28 @@
 import { friendBucketAssignments, friendConnections, keyGrants } from "@pluralscape/db/pg";
-import { brandId, now, toUnixMillis } from "@pluralscape/types";
-import { and, desc, eq, inArray, isNull, lt, or, sql } from "drizzle-orm";
+import { brandId, now } from "@pluralscape/types";
+import { and, eq, inArray, isNull, sql } from "drizzle-orm";
 
-import { HTTP_CONFLICT, HTTP_NOT_FOUND } from "../http.constants.js";
-import { assertAccountOwnership } from "../lib/account-ownership.js";
-import { ApiHttpError } from "../lib/api-error.js";
-import { encryptedBlobToBase64OrNull, validateEncryptedBlob } from "../lib/encrypted-blob.js";
-import { archiveAccountEntity, restoreAccountEntity } from "../lib/entity-lifecycle.js";
-import { assertOccUpdated } from "../lib/occ-update.js";
-import { buildCompositePaginatedResult, fromCompositeCursor } from "../lib/pagination.js";
-import {
-  withAccountRead,
-  withAccountTransaction,
-  withCrossAccountTransaction,
-} from "../lib/rls-context.js";
-import {
-  DEFAULT_PAGE_LIMIT,
-  MAX_ENCRYPTED_DATA_BYTES,
-  MAX_PAGE_LIMIT,
-} from "../service.constants.js";
+import { HTTP_CONFLICT, HTTP_NOT_FOUND } from "../../http.constants.js";
+import { assertAccountOwnership } from "../../lib/account-ownership.js";
+import { ApiHttpError } from "../../lib/api-error.js";
+import { withCrossAccountTransaction } from "../../lib/rls-context.js";
+import { dispatchWebhookEvent } from "../webhook-dispatcher.js";
 
-import { dispatchWebhookEvent } from "./webhook-dispatcher.js";
+import { toFriendConnectionResult } from "./internal.js";
 
-import type { AuditWriter } from "../lib/audit-writer.js";
-import type { AuthContext } from "../lib/auth-context.js";
-import type { AccountArchivableEntityConfig } from "../lib/entity-lifecycle.js";
+import type { FriendConnectionResult, FriendConnectionWithRotations } from "./internal.js";
+import type { AuditWriter } from "../../lib/audit-writer.js";
+import type { AuthContext } from "../../lib/auth-context.js";
 import type {
   AccountId,
   AuditEventType,
   BucketId,
   FriendConnectionId,
   FriendConnectionStatus,
-  PaginatedResult,
   SystemId,
   UnixMillis,
 } from "@pluralscape/types";
 import type { PostgresJsDatabase } from "drizzle-orm/postgres-js";
-
-// ── Types ───────────────────────────────────────────────────────────
-
-export interface FriendConnectionResult {
-  readonly id: FriendConnectionId;
-  readonly accountId: AccountId;
-  readonly friendAccountId: AccountId;
-  readonly status: FriendConnectionStatus;
-  readonly encryptedData: string | null;
-  readonly version: number;
-  readonly createdAt: UnixMillis;
-  readonly updatedAt: UnixMillis;
-}
-
-export interface FriendConnectionWithRotations extends FriendConnectionResult {
-  readonly pendingRotations: ReadonlyArray<{
-    readonly systemId: SystemId;
-    readonly bucketId: BucketId;
-  }>;
-}
-
-interface ListFriendConnectionOpts {
-  readonly cursor?: string;
-  readonly limit?: number;
-  readonly includeArchived?: boolean;
-  readonly status?: FriendConnectionStatus;
-}
-
-interface UpdateFriendVisibilityParams {
-  readonly encryptedData: string;
-  readonly version: number;
-}
 
 // ── Constants ───────────────────────────────────────────────────────
 
@@ -87,110 +42,6 @@ const AUDIT_FRIEND_ACCEPTED: AuditEventType = "friend-connection.accepted";
 const AUDIT_FRIEND_REJECTED: AuditEventType = "friend-connection.rejected";
 const AUDIT_FRIEND_BLOCKED: AuditEventType = "friend-connection.blocked";
 const AUDIT_FRIEND_REMOVED: AuditEventType = "friend-connection.removed";
-const AUDIT_VISIBILITY_UPDATED: AuditEventType = "friend-visibility.updated";
-const AUDIT_FRIEND_ARCHIVED: AuditEventType = "friend-connection.archived";
-const AUDIT_FRIEND_RESTORED: AuditEventType = "friend-connection.restored";
-
-// ── Helpers ─────────────────────────────────────────────────────────
-
-function toFriendConnectionResult(
-  row: typeof friendConnections.$inferSelect,
-): FriendConnectionResult {
-  return {
-    id: brandId<FriendConnectionId>(row.id),
-    accountId: brandId<AccountId>(row.accountId),
-    friendAccountId: brandId<AccountId>(row.friendAccountId),
-    status: row.status,
-    encryptedData: encryptedBlobToBase64OrNull(row.encryptedData),
-    version: row.version,
-    createdAt: toUnixMillis(row.createdAt),
-    updatedAt: toUnixMillis(row.updatedAt),
-  };
-}
-
-// ── LIST ────────────────────────────────────────────────────────────
-
-export async function listFriendConnections(
-  db: PostgresJsDatabase,
-  accountId: AccountId,
-  auth: AuthContext,
-  opts: ListFriendConnectionOpts = {},
-): Promise<PaginatedResult<FriendConnectionResult>> {
-  assertAccountOwnership(accountId, auth);
-
-  const effectiveLimit = Math.min(opts.limit ?? DEFAULT_PAGE_LIMIT, MAX_PAGE_LIMIT);
-
-  return withAccountRead(db, accountId, async (tx) => {
-    const conditions = [eq(friendConnections.accountId, accountId)];
-
-    if (!opts.includeArchived) {
-      conditions.push(eq(friendConnections.archived, false));
-    }
-
-    if (opts.status) {
-      conditions.push(eq(friendConnections.status, opts.status));
-    }
-
-    if (opts.cursor) {
-      const decoded = fromCompositeCursor(opts.cursor, "friend-connection");
-      const cursorCondition = or(
-        lt(friendConnections.createdAt, decoded.sortValue),
-        and(
-          eq(friendConnections.createdAt, decoded.sortValue),
-          lt(friendConnections.id, decoded.id),
-        ),
-      );
-      if (cursorCondition) {
-        conditions.push(cursorCondition);
-      }
-    }
-
-    const rows = await tx
-      .select()
-      .from(friendConnections)
-      .where(and(...conditions))
-      .orderBy(desc(friendConnections.createdAt), desc(friendConnections.id))
-      .limit(effectiveLimit + 1);
-
-    return buildCompositePaginatedResult(
-      rows,
-      effectiveLimit,
-      toFriendConnectionResult,
-      (i) => i.createdAt,
-    );
-  });
-}
-
-// ── GET ─────────────────────────────────────────────────────────────
-
-export async function getFriendConnection(
-  db: PostgresJsDatabase,
-  accountId: AccountId,
-  connectionId: FriendConnectionId,
-  auth: AuthContext,
-): Promise<FriendConnectionResult> {
-  assertAccountOwnership(accountId, auth);
-
-  return withAccountRead(db, accountId, async (tx) => {
-    const [row] = await tx
-      .select()
-      .from(friendConnections)
-      .where(
-        and(
-          eq(friendConnections.id, connectionId),
-          eq(friendConnections.accountId, accountId),
-          eq(friendConnections.archived, false),
-        ),
-      )
-      .limit(1);
-
-    if (!row) {
-      throw new ApiHttpError(HTTP_NOT_FOUND, "NOT_FOUND", "Friend connection not found");
-    }
-
-    return toFriendConnectionResult(row);
-  });
-}
 
 // ── Private helpers ─────────────────────────────────────────────────
 
@@ -321,8 +172,6 @@ async function cleanupBucketAssignments(
     bucketId: brandId<BucketId>(a.bucketId),
   }));
 }
-
-// ── TERMINATE (block / remove) ─────────────────────────────────────
 
 /**
  * Shared implementation for both block and remove operations.
@@ -522,113 +371,4 @@ export function removeFriendConnection(
     auditEventType: AUDIT_FRIEND_REMOVED,
     auditDetail: "Friend connection removed",
   });
-}
-
-// ── UPDATE VISIBILITY ───────────────────────────────────────────────
-
-export async function updateFriendVisibility(
-  db: PostgresJsDatabase,
-  accountId: AccountId,
-  connectionId: FriendConnectionId,
-  params: UpdateFriendVisibilityParams,
-  auth: AuthContext,
-  audit: AuditWriter,
-): Promise<FriendConnectionResult> {
-  assertAccountOwnership(accountId, auth);
-
-  const blob = validateEncryptedBlob(params.encryptedData, MAX_ENCRYPTED_DATA_BYTES);
-  const version = params.version;
-  const timestamp = now();
-
-  return withAccountTransaction(db, accountId, async (tx) => {
-    const updated = await tx
-      .update(friendConnections)
-      .set({
-        encryptedData: blob,
-        updatedAt: timestamp,
-        version: sql`${friendConnections.version} + 1`,
-      })
-      .where(
-        and(
-          eq(friendConnections.id, connectionId),
-          eq(friendConnections.accountId, accountId),
-          eq(friendConnections.version, version),
-          eq(friendConnections.archived, false),
-        ),
-      )
-      .returning();
-
-    const row = await assertOccUpdated(
-      updated,
-      async () => {
-        const [existing] = await tx
-          .select({ id: friendConnections.id })
-          .from(friendConnections)
-          .where(
-            and(eq(friendConnections.id, connectionId), eq(friendConnections.accountId, accountId)),
-          )
-          .limit(1);
-        return existing;
-      },
-      "Friend connection",
-    );
-
-    await audit(tx, {
-      eventType: AUDIT_VISIBILITY_UPDATED,
-      actor: { kind: "account", id: auth.accountId },
-      detail: "Friend visibility updated",
-      accountId,
-    });
-
-    return toFriendConnectionResult(row);
-  });
-}
-
-// ── Entity lifecycle config (account-scoped) ──────────────────────
-
-const FRIEND_CONNECTION_LIFECYCLE: AccountArchivableEntityConfig<FriendConnectionId> = {
-  table: friendConnections,
-  columns: {
-    id: friendConnections.id,
-    accountId: friendConnections.accountId,
-    archived: friendConnections.archived,
-    archivedAt: friendConnections.archivedAt,
-    updatedAt: friendConnections.updatedAt,
-    version: friendConnections.version,
-  },
-  entityName: "Friend connection",
-  archiveEvent: AUDIT_FRIEND_ARCHIVED,
-  restoreEvent: AUDIT_FRIEND_RESTORED,
-};
-
-// ── ARCHIVE ─────────────────────────────────────────────────────────
-
-export async function archiveFriendConnection(
-  db: PostgresJsDatabase,
-  accountId: AccountId,
-  connectionId: FriendConnectionId,
-  auth: AuthContext,
-  audit: AuditWriter,
-): Promise<void> {
-  await archiveAccountEntity(db, accountId, connectionId, auth, audit, FRIEND_CONNECTION_LIFECYCLE);
-}
-
-// ── RESTORE ─────────────────────────────────────────────────────────
-
-export async function restoreFriendConnection(
-  db: PostgresJsDatabase,
-  accountId: AccountId,
-  connectionId: FriendConnectionId,
-  auth: AuthContext,
-  audit: AuditWriter,
-): Promise<FriendConnectionResult> {
-  return restoreAccountEntity(
-    db,
-    accountId,
-    connectionId,
-    auth,
-    audit,
-    FRIEND_CONNECTION_LIFECYCLE,
-    (row) => toFriendConnectionResult(row as typeof friendConnections.$inferSelect),
-  );
 }
